refactor(attester): add SNP attester and guest ioctl layouts
Introduce attestation::attester as the user-facing API for guest
attestation ioctls. Make firmware::guest an internal /dev/sev-guest
transport layer (ioclt layouts, cert table, types).
