--- src/firmware/guest/types/mod.rs
@@ -1,7 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-
-#[cfg(feature = "snp")]
-mod snp;
-
-#[cfg(feature = "snp")]
-pub use self::snp::*;
