+++ src/firmware/linux/mod.rs
@@ -2,7 +2,4 @@
 
 pub mod host;
 
-#[cfg(feature = "snp")]
-pub mod guest;
-
 pub(crate) const _4K_PAGE: usize = 4096;
